Changed all id type to long from int.

+++ src/main/kotlin/onlytrade/app/product/data/table/ProductTable.kt
@@ -4,8 +4,8 @@ import org.jetbrains.exposed.dao.id.IntIdTable
 
 
 object ProductTable : IntIdTable(name = "product") {
-    val userId = integer("userId")
-    val subcategoryId = integer("subcategoryId")
+    val userId = long("userId")
+    val subcategoryId = long("subcategoryId")
     val name = text("name")
     val description = text("desc")
     val estPrice = double("estPrice")
